requirements.txt
Django==3.2.6
djangorestframework==3.12.4
psycopg2==2.8.6
django-cors-headers==3.8.0
drf-yasg==1.20.0
django-url-filter==0.3.15
django-cors-headers==3.8.0
